# 3-technical/3.1-system-foundation/architecture/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AKAIUNSAN Attendance + Payroll API
  version: 1.0.0
  description: |
    REST API for AKAIUNSAN attendance and payroll systems (PRD-EPIC-002, Phase 0).

    **Auth:** Bearer JWT (15-min TTL) + refresh token cookie (30 days).
    **Internal:** `X-Internal-API-Key` header (service-to-service only).
    **Rate limits:** Production API instances enforce a global 100 requests/minute limit; auth endpoints add OTP/challenge abuse controls.
    **Time:** ISO 8601 UTC; VN time (UTC+7) at presentation layer.
    **Money:** VNĐ as string (Decimal); never float.
    **Errors:** `{ error: { code, message, details? } }`

    Full schemas live in this file. Full business rules in
    [Domain Specs](../domain-specs.md). Architecture in
    [System Design](../../design-standards/system-design.md).

    Code conventions in [Coding Standards](../../design-standards/coding-standards.md).

  contact:
    name: AKAIUNSAN Engineering
    email: engineering@ak.local
...
security:
  - bearerAuth: []
paths:
...
  /v1/projects:
    get:
      tags: [projects]
      summary: List projects
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, paused, ended] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
  /v1/projects/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [projects]
      summary: Get project
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
    patch:
      tags: [projects]
      summary: Update project (BO/system admin; permission-gated)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{id}/supervisors:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [projects]
      summary: List explicit project supervisor memberships (BO/system admin only)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectSupervisor' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [projects]
      summary: Grant explicit project supervisor membership (BO/system admin only, audited)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '200':
          description: Membership granted or already present
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectSupervisor' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/projects/{id}/supervisors/{userId}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: userId
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [projects]
      summary: Revoke project supervisor membership immediately (BO/system admin only, audited)
      responses:
        '204': { description: Membership revoked }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===== SHIFTS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (15-min TTL)
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Shared secret, rotated quarterly
...
  responses:
...
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Project:
      type: object
      required: [id, code, name, clientName, address, latitude, longitude, geofenceRadiusMeters, status]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PRJ001 }
        name: { type: string }
        clientName: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        geofenceRadiusMeters: { type: integer, default: 100 }
        contractStartDate: { type: string, format: date }
        contractEndDate: { type: [string, 'null'], format: date }
        status: { type: string, enum: [active, paused, ended] }
    ProjectCreate:
      type: object
      required: [code, name, clientName, address, latitude, longitude, contractStartDate]
      properties:
        code: { type: string }
        name: { type: string }
        clientName: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        geofenceRadiusMeters: { type: integer, default: 100 }
        contractStartDate: { type: string, format: date }
        contractEndDate: { type: string, format: date }
    ProjectUpdate:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        clientName: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        geofenceRadiusMeters: { type: integer, minimum: 1 }
        contractStartDate: { type: string, format: date }
        contractEndDate: { type: string, format: date }
        status: { type: string, enum: [active, paused, ended] }
        reportTemplateConfig: { type: object, additionalProperties: true }
    ProjectSupervisor:
      type: object
      required: [userId, createdAt]
      properties:
        projectId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        assignedById: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        supervisor:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: [string, 'null'], format: email }
            phone: { type: string }
            status: { type: string, enum: [active, inactive, suspended] }

    # ===== Shift =====
...